backend/pyproject.toml
[project]
name = "akuwallet-backend"
version = "0.1.0"
description = "Python FastAPI backend for Wallet personal finance app"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.135.1",
    # "supabase==2.28.0", # TEMP: disabled while using local SQLite
    "pydantic==2.12.5",
    "google-genai>=1.17.0",
    "aiosqlite==0.22.1",
    "python-dotenv>=1.0.0",
]
